+++ src/constants.ts
@@ -83,6 +83,8 @@ export const TEST_IDS = {
     fieldDropdownAlignOption: createSelector((name: unknown) => `link-editor field-dropdown-align-${name}`),
     fieldDropdownButtonSize: createSelector('data-testid link-editor field-dropdown-button-size'),
     fieldDropdownButtonSizeOption: createSelector((name: unknown) => `link-editor field-dropdown-button-size-${name}`),
+    fieldShowCustomIcon: createSelector('data-testid link-editor field-show-custom-icon'),
+    fieldCustomIconUrl: createSelector('data-testid link-editor field-custom-icon-url'),
     fieldAlignContent: createSelector('data-testid link-editor field-align-content'),
     fieldAlignContentOption: createSelector((name: unknown) => `link-editor field-align-content-${name}`),
   },