+++ src/types/panel.ts
@@ -311,6 +311,20 @@ export interface LinkConfig {
   dropdownConfig?: DropdownConfig;
 
   /**
+   * Use custom images
+   *
+   * @type {boolean}
+   */
+  showCustomIcons?: boolean;
+
+  /**
+   * Custom image url
+   *
+   * @type {string}
+   */
+  customIconUrl?: string;
+
+  /**
    * Align content position
    *
    * @type {AlignContentPositionType}